+++ fNet-common/src/main/java/com/fnet/common/codec/MessageEncoder.java
@@ -12,7 +12,7 @@ public class MessageEncoder extends MessageToByteEncoder<Message> {
 
     @Override
     protected void encode(ChannelHandlerContext ctx, Message msg, ByteBuf out) {
-        out.writeInt(msg.getType().getCode());
+        out.writeByte(msg.getType().getCode());
         out.writeInt(msg.getOuterChannelId());
         if (msg.getData() == null) {
             out.writeInt(0);
